fix but with shop items
